raster: separate raster_persist and raster_stroke
